+++ tests/manual/wal_explorer.cpp
@@ -49,6 +49,10 @@ std::string StateDeltaTypeToString(database::StateDelta::Type type) {
       return "BUILD_EXISTENCE_CONSTRAINT";
     case database::StateDelta::Type::DROP_EXISTENCE_CONSTRAINT:
       return "DROP_EXISTENCE_CONSTRAINT";
+    case database::StateDelta::Type::BUILD_UNIQUE_CONSTRAINT:
+      return "BUILD_UNIQUE_CONSTRAINT";
+    case database::StateDelta::Type::DROP_UNIQUE_CONSTRAINT:
+      return "DROP_UNIQUE_CONSTRAINT";
   }
 }